Bug 1245776 - part3: import shared/test/head in styleeditor/test/head;r=gl
5 tests of the styleeditor suite are importing shared/test/head.
Now that this also includes shared-head.js, it conflicts with the existing
styleeditor head file.

For now, the import was moved to the head, so it's always imported.
The tests have been updated accordingly.

Should file a follow-up to split the inspector test suites head files in a
head.js + helpers.js (see commandline/test). It should be possible to
include helpers in other tests without impact on the test lifecycle.