Bug 1260711 - Fix error 'IndexSizeError: Index or size is negative or greater than the allowed amount' when using the animation inspector. r=pbro
Error was raised in drawGraphElementBackground when called with the graphView
being `0` This could happen because in animation-timeline.js , there is a
resize event handler that could be fired while the animation panel width is 0.
We fix this by making sure the animation panel does not have a width of zero
in the resize handler.