Bug 1248459 - Don't query out-of-bounds selection; r=masayuki
During a query selection event, fail if the selection is outside of the
editor's root content. This can happen if the placeholder text in an
input field is somehow selected. The placeholder is in a separate
element outside of the root content.

Also fix a bug in test_imestate, where the selection was not properly
reset at the start of a test.