Bug 1257874 - Use getProperties instead of getFrames in animation-detail. r=pbro
Add a getProperties function to the animation actor to map KeyframeEffectReadOnly.getProperties
Call this new function in animation-detail and adapt the code as the structure
of the returned object structure is different from what getFrames returns.
Adapt a couple tests to the new object structure client-side, and a test server-side
to check getProperties returns what it should.