Bug 1228655 - Move TabsInTitlebar to its own file to remove ifdefs. r=Gijs
TabsInTitlebar has various ifdefs to avoid shipping code that's not gonna be used on Linux. Instead of complicating the logic with AppConstants checks and shipping this extra code unecessarily on Linux, i've separated the code to a different file that can be replaced by a stub implementation.

RM: Adapted, kept old AppMenu code and removed CustomizeUI lisitener, or menus would double on Mac