Bug 1228673 - Clean-up 'Then' before ~MediaTimer. r=jya
Destroying an unlapsed timer will trigger an assertion in its 'Then' request.
So in case the Fuzzing Wrapper is flushed or errors when there are pending
delayed frames, the 'Then' request attached to the MediaTimer is now
disconnected before being let go.
This also needs to be done when shutting down.