Bug 1272964: [MSE] P5. Default to skipping to the next keyframe if no keyframe was found past currentTime.
If no keyframe are found after our time threshold, we can still skip to another keyframe (despite being prior the desired time).
So this is just a workaround for our inability to tell the MDSM when to enter buffering mode and instead the MDSM incorrectly uses the time of the last frame returned.