Bug 1136226 - Remove SimdLane enumeration. r=sunfish
We're about to add 16-lane and 8-lane SIMD types to IonMonkey, so we don't want
an enum to identify lanes. Just use an unsigned instead.

Also note that SIMD.js no longer has the .x .y .z .w properties.

Change a few lane < 4 checks to use the number of lanes in the Simd type.