Bug 1263197 - part 3 - remove GetUserMediaLog function from MediaEngineWebRTC.cpp; r=jesup
GetUserMediaLog is a holdover from when when used PRLogModule and needed
to avoid a static constructor.  Now we have LazyLogModule, which doesn't
invoke static constructors, and so we can use that directly, rather than
indirecting through GetUserMediaLog.  This change also helps avoid an
unused function warning when logging is disabled.