Bug 1276565 - Clamp out-of-range tab selection shortcuts and indexes instead of ignoring them. r=Gijs r=dao
CMD + a number key is a keyboard shortcut to select that tab number. (CMD+9 will always select the right-most tab, regardless of the tab count.) Currently, if the number is greater than the tab count, then the keyboard shortcut is ignored. With this patch, the right-most tab will be selected instead. For example, if there are five tabs and the user accidentally enters CMD+6 instead of CMD+5, selecting tab #5 is probably what they wanted instead of ignoring the keyboard shortcut.

Also expand the tab selection tests to cover more edge cases.