Bug 1233927 - Switching between large JS files in debugger is slow. r=jlong
Changes the way that source files are loaded in the editor. Previously,
source text and mode were set on the existing editor document.

Now, source files are their own CodeMirror Documents, so when it comes time
to showing them in the editor, it's just a matter of swapping one
document for another.

Notes:
+ The DebuggerView now has a _setEditorDocument method for showing a
source document
+ The Editor now has support for creating documents and replacing
documents.