Bug 1274520 part 2 - Listen control events in video controls on system group. r=gijs,jaws
Listeners in the system group are invoked after listeners in the default
group, so we can check defaultPrevented synchronously, and we have to
change the play state according to the state of the play button.

Due to the same reason above, we need to check the UI async in tests.