Bug 1264822 part 3 - Add a test that non-animatable properties are not accessed; r=hiro
We already have a test under
web-animations/animation-model/animation-types/not-animatable.html that checks
that non-animatable properties don't show up in the result from getKeyframes().
However, the spec says we should not even *read* those properties off the input
object(s).

This patch adds a test to check that we don't read non-animatable properties.
In a later patch we will remove the existing test since it will be redundant
after this test is added.