Bug 1278403 - WebGL doesn't need validation. - r=jrmuizel
We do all validation in our WebGL layer, we only really generate GL_OOM
from the driver, no other errors.

This is opposed to SkiaGL, which may or may not rely on errors returned
from the driver. This is why I am not just turning on ABORT_ON_ERROR for
all GLContexts.

(Also, later, we should be creating no-error contexts on non-DEBUG builds,
so the driver doesn't replicate our validation. ANGLE already supports
this)